fix: utils updated with toQueryParams
