+++ src/utils.ts
@@ -2,6 +2,7 @@ import { OPEN_AI_API_KEY, PORTKEY_HEADER_PREFIX } from "./constants";
 import { createResponseHeaders } from "./streaming";
 import OpenAI from "openai";
 import type { Portkey } from "./index";
+import { UserInviteListParams, UsersListParams, WorkspaceMemberListParams, WorkspacesListParams } from "./apis/admin";
 
 type PlatformProperties = {
 	"x-portkey-runtime"?: string,
@@ -129,4 +130,15 @@ export function initOpenAIClient(client: Portkey){
 		defaultHeaders: defaultHeadersBuilder(client),
 		maxRetries: 0
 	})
+}
+export function toQueryParams(params?: (UsersListParams | UserInviteListParams | WorkspacesListParams | WorkspaceMemberListParams)): string {
+    if (!params) {
+        return '';
+    }
+    const queryParams = Object.entries(params)
+        .filter(([, value]) => value !== undefined && value !== null)
+        .map(([key, value]) => `${key}=${value}`)
+        .join('&');
+    
+    return queryParams ? `?${queryParams}` : '';
 }
